Migrate netutil methods into /utils/net.go
